Bulk update of recent changes in our frameworks for iOS 4.2 and other updates.
Highlights:

Configurations:
  - Switch from llvm-gcc-4.2 to clang for both iOS and Mac.
  - Twiddle our debug information to have what we need for local symbolication
  - Link against the latest iphoneos, but target 3.2 as the minimum.

Overall:
  - Fixed new warnings reported by clang top-of-tree.
  - Run genstrings at build time in various places.
  - Various iOS 4.2 updates.
  - Optimized images.

OmniBase:
  - Adding some inttypes-style format macros for Apple-defined types
  - Catch common error of key-specific variant of
    +automaticallyNotifiesObserversForKey: using ...ForFoo instead of
    ...OfFoo.

OmniFoundation:
  - Extensions and clean up of NSUndoManager debugging extension.
  - Added a copy of Apple's NetReachability so we could avoid adding it in
    multiple apps.

OmniFileStore:
  - Brought over HTTP redirection fixes from OmniFocus/1.x
  - Added an iOS target.
  - Added options for directory enumeration
  - Report the last modification date in OFSFileInfo entries.
  - Cleaned up API for asynchronous operations to be more consistent with
    synchronous.
    - Async methods always return a non-nil operation on success.
    - Async operations have a protocol declared, including a startup method
      (so you can store them in an ivar w/o worrying about a race with
      completion).
  - Added async write operations with support for tracking progress.
  - Updated the 'ofs' command line tool with a new 'acp' command to test async
    read/write and progress.

OmniAppKit:
  - Various binding fixes for OAPopupDatePicker.
  - Fixed issues with OAFontDescriptor's lookup of symbolic traits;
    consistently fall back to the internal CTFontRef's values if the attribute
    dictionary doesn't specify a value.
  - Add more {NS,CT} text attribute mappings in OATextAttributes.[hm].
  - Fixed some OATargetSelection behavior with sheets.
  - OAScriptToolbarHelper retains its toolbar item for the duration of the
    script to avoid a zombie if the script closes the containing window.

OmniQuartz:
  - Fix OQLayerRemovalAnimation to call the completion handler if we're not
    animating.
  - Fixed some glitches with OQHoleLayer.
  - Be more careful about colorspace conversions in OQColor. Sure wish iOS had
    a linear RGB colorspace.

OmniInspector:
  - Some cleanup for single-inspector configurations

OmniSoftwareUpdate:
  - Turned on signature checking.
  - Added an iOS target that just sends OS/hardware info (just the analytics
    part), along with hooks for integrating into apps.
    - This isn't enabled in any of our apps yet since we don't have UI for it
      we like yet.
  - Do the checks in the main executable; the OmniSoftwareUpdateCheck tool
    still exists for debugging, but we can't use it on iOS so the framework no
    longer uses it.
  - Fixed some leaks now that we do OSU checks in the main process instead of
    letting them die with the background tool.
  - Use OFNetReachability instead of having our own reachability code.

OmniUnzip:
  - Implemented a convenience method +createZipFile:fromFiles:error:, for
  	easily zipping up any number of files or directories into a single
  	.zip wrapper.  (Handy for attaching to an email message or uploading
  	to a WebDAV server.)

OmniUI:
  - Added WebDAV-based import/export support for documents.
  - Added iWork-like export sheet.
  - Fixes/extensions to OUIUndoBarButtonItem to help avoid multiple popovers.
  - Added an optional scrubber on the document picker to quickly navigate
    the list of documents.
  - Add support for sorting documents by name or modification date.
  - Added OUIAnimationSequence for running a sequence of block-based
    animations.
  - Added OUIWithoutAnimating(), taking an animation block and running it with
    +[UIView areAnimationsEnabled] set to NO.
  - Add hooks for printing.
  - Add initial pass at OmniSoftwareUpdate analytics support. Disabled for now
    until we decide we decide on good UI (or fail to do so).
  - Added debugging checks that calls to -convert{Point,Rect}:{to,from}View:
    are not made with views without a common ancestor.
  - Text selection tweaks:
	  - Changed the color of the highlight and carets.
    - Replaced the drawn knob with a .png.
    - Round off the translation portion of the loupe zoom transform to reduce
      crawl.
		- Positioning the rectangular loupe based on the text selection's
		  endpoint's location, rather than the location of the touch.
  - Update font inspector to work more like iWork. Strikethrough support is
    optional since we haven't yet implemented it in our text rendering.
  - Add support for underline-by-word as a render-time transform.

+++ Frameworks/OmniAppKit/OAUtilities.h
@@ -10,11 +10,12 @@
 #import <OmniFoundation/NSNumber-OFExtensions-CGTypes.h>
 
 #import <math.h>
-#include <tgmath.h>
+#import <tgmath.h>
 
+// Used to have a ppc frsqrte version, but don't have an x86_64 version right now
 static inline CGFloat OAFastReciprocalSquareRoot(CGFloat x)
 {
-    return 1.0f / sqrt(x);
+    return (CGFloat)(1.0f / sqrt(x));
 }
 
 #if defined(__COREGRAPHICS__) && !defined(__cplusplus)
